refactor(geode-core): extract backend, traits, prelude modules with deprecated root shims
Move the backend-selection machinery and device smoke surface out of
lib.rs into a directory-backed `backend` module, extract the core
`Mesh`/`Element`/`Operator` traits into a `traits` module, and add an
initial `prelude`. Pure refactor — no behavior change.

- `backend/mod.rs` now owns both `compile_error!` guards and both
  `std::cfg_select!` cascades (`DefaultBackend`, private `BACKEND_NAME`),
  plus `DeviceInfo`, `device_info`, private `default_device_label`,
  `smoke_add`, and the relocated `#[cfg(test)] mod tests`. `pub mod
  backend;` is declared unconditionally in lib.rs so the "no backend"
  guard still fires; the cfg predicates resolve identically from the
  submodule.
- `traits/mod.rs` holds the three traits; they stay reachable at the
  crate root via a non-deprecated re-export (preserves the
  `crate::Mesh` intra-doc link) and are added to the prelude.
- `prelude.rs` re-exports the four public backend items + three traits
  from their canonical module paths (warning-free under -D warnings).
- Four genuinely-public moved items (`DefaultBackend`, `DeviceInfo`,
  `device_info`, `smoke_add`) get `#[deprecated]` root `pub use` shims;
  the private `BACKEND_NAME`/`default_device_label` move silently.
- Migrate in-crate callers off the deprecated root path: driven.rs and
  sparse.rs test modules, the fe_assemble.rs doctest, and the geode-cli
  `smoke_add` call.

Part of epic #377.
