+++ .loom-managed
@@ -1,6 +1,6 @@
 # Loom-managed worktree marker
 # Created by .loom/scripts/worktree.sh
-# Issue: 354
-# Branch: feature/issue-354
+# Issue: 378
+# Branch: feature/issue-378
 # Removing this file makes Loom treat the worktree as user-owned and refuse
 # to clean it up automatically.
